+++ frontend/src/App.jsx
@@ -1,13 +1,13 @@
 import { Route, Routes } from "react-router-dom";
 import Header from "./components/Header";
 import CategoryBar from "./components/CategoryBar";
-import CategoryPage from "./components/CategoryPage";
+import CategoryPage from "./pages/CategoryPage";
 import Footer from "./components/Footer";
-import Search from "./components/Search";
+import Search from "./pages/Search";
 import Register from "./pages/Register";
 import Login from "./pages/Login";
 import ForgetPass from "./pages/ForgetPass";
-import Home from "./components/Home";
+import Home from "./pages/Home";
 import "./App.css";
 
 function App() {
