+++ frontend/src/pages/CategoryPage.jsx
